Fixed quotation price issue

+++ database/migrations/2021_04_27_210751_alter_quotation_requests_table_v1.php
@@ -14,7 +14,7 @@ class AlterQuotationRequestsTableV1 extends Migration
     public function up()
     {
         Schema::table('quotation_requests', function (Blueprint $table) {
-            $table->double('quotation_amount', 15, 2)->nullable()->change();
+            $table->decimal('quotation_amount', 15, 2)->nullable()->change();
         });
     }
