+++ database/migrations/2021_04_27_211022_alter_quotation_request_details_table_v1.php
@@ -14,7 +14,7 @@ class AlterQuotationRequestDetailsTableV1 extends Migration
     public function up()
     {
         Schema::table('quotation_request_details', function (Blueprint $table) {
-            $table->double('price', 15, 2)->change();
+            $table->decimal('price', 15, 2)->change();
         });
     }
